Perform exponentiation in normalized space
By refactoring the code slightly, we are able to perform exponentiation in
the normalized space of the reciprocal. This saves two shifts in each mod
operation, as well as a few modulus operations that are not necessary in the
normalized space.

We can also use the reciprocal for the division after the term is computed,
which saves some time and helps amortize the cost of computing the reciprocal.